[ORC][JITLink] Factor out common VTune data structures into OrcShared
Fixed the cyclic library dependency: https://github.com/llvm/llvm-project/pull/83909
